requirements.txt
annotated-types==0.5.0
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.7.1
Django==4.2.5
django-cors-headers==4.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-spectacular==0.26.5
flake8==6.1.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
mccabe==0.7.0
oauthlib==3.2.2
packaging==23.2
pluggy==1.3.0
psycopg2==2.9.7
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
PyJWT==2.8.0
pytest==7.4.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.0
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.6
ruff==0.0.292
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
StrEnum==0.4.15
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
